requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchtext>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# For data processing
tqdm>=4.62.0
nltk>=3.6.0
pyyaml>=5.4.0

# For development
jupyter>=1.0.0
ipykernel>=6.0.0
black>=21.7b0
flake8>=3.9.0

# For visualization
opencv-python>=4.5.0

# For model training
tensorboard>=2.6.0

# For web interface (optional)
flask>=2.0.0